% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTCA.R
\name{sprtca}
\alias{sprtca}
\title{Semi-parametric randomization test for continuous association}
\usage{
sprtca(iter, character.var, numeric.var, net)
}
\arguments{
\item{iter}{Number of iterations for randomization}

\item{character.var}{Variable 1}

\item{numeric.var}{Variable 2}

\item{net}{A  RDS data frame.}
}
\description{
SPRTCA was developed to infer bivariate association between two variables
collected using respondent-driven sampling (RDS) when one or both variables are continuous.
}
